% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jul_con.R
\name{jul_con}
\alias{jul_con}
\title{Turns Month/Day/Year into Julian Date}
\usage{
jul_con(x)
}
\arguments{
\item{data.frame}{three columns (month/day/year)}
}
\description{
This function lets you convert regular dates to julian ones
}
\examples{
jul_con()
}
\keyword{date}
\keyword{julian}

