% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mround.R
\name{mround}
\alias{mround}
\title{Rounds numbers into categories by dec}
\usage{
mround(x, dec)
}
\arguments{
\item{a}{number that you want to round to a dec (decimal)}
}
\description{
This function lets you convert regular dates to ordinal ones
}
\examples{
mround()
}
\keyword{rounding}

