% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_date_con.R
\name{ordinal_date_con}
\alias{ordinal_date_con}
\title{Turns Month/Day/Year into Ordinal Date}
\usage{
ordinal_date_con(x)
}
\arguments{
\item{data.frame}{three columns (month/day/year)}
}
\description{
This function lets you convert regular dates to ordinal ones
}
\examples{
ordinal_date_con()
}
\keyword{date}
\keyword{ordinal}
