\name{print.rel}
\alias{print.rel}
\title{Class print structure}
\description{The structure of the printed output for package functions}
\usage{
  \method{print}{rel}(x, ...)
}
\arguments{
  \item{x}{rel function output}
  \item{...}{Other}
}
\value{The printed output is detailed in the description of the respective functions}