% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/cat_function.R
\name{rail_function}
\alias{rail_function}
\title{A Rail Function}
\usage{
rail_function(love = TRUE)
}
\arguments{
\item{love}{Do you love rails? (you SHOULD!) Defaults to TRUE.}
}
\description{
This function allows you to express your love of rails.
Unsure what rails are? https://en.wikipedia.org/wiki/Rallidae
}
\examples{
rail_function()
}
\keyword{rails}

