\name{rel-package}
\alias{rel}
\alias{rel-package}
\docType{package}
\title{Various reliability coefficients}
\description{
  \code{rel} provides functions for evaluating agreement between measurements: Bennett et als S (\code{\link{bags}}), Cohen's kappa and weighted kappa (\code{\link{ckap}}), Gwet's AC1 and AC2 (\code{\link{gac}}), Krippendorff's alpha (\code{\link{kra}}), and Scott's pi (\code{\link{spi}}).
}
\keyword{ 
  package
  univar
}