% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_by_ten.R
\name{round_by_10}
\alias{round_by_10}
\title{round values by 10}
\usage{
round_by_10(x)
}
\description{
round values by 10
}
\examples{
>round_by_10(17)
>20
}
\keyword{10}
\keyword{by}
\keyword{numeric}
\keyword{round}
\keyword{values}

