% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale_x_ordinaldate.R
\name{scale_x_ordinaldate}
\alias{scale_x_ordinaldate}
\title{go from ordinal dates to readable dates in axis}
\usage{
scale_x_ordinaldate(year)
}
\description{
go from ordinal dates to readable dates in axis
}
\examples{
ggplot(data=dat, aes(x=x, y=y))+geom_point()+
}
\keyword{ggplot2}

