% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_oridinaldate.R
\name{scale_y_ordinaldate}
\alias{scale_y_ordinaldate}
\title{go from ordinal dates to readable dates in axis}
\usage{
scale_y_ordinaldate(year)
}
\description{
go from ordinal dates to readable dates in axis
}
\examples{
ggplot(data=dat, aes(x=x, y=y))+geom_point()+
}
\keyword{ggplot2}

