% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{smoother}
\alias{smoother}
\title{smoother spline}
\usage{
smoother(x)
}
\arguments{
\item{input}{is melted data.frame with columns (jdate, variable and value)}
}
\description{
wrapper for smooth_spline makes it into the format I graph from
}
\examples{
smoother()
}
\keyword{smooth}
\keyword{spline}

