% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sora_spline.R
\name{sora_spline}
\alias{sora_spline}
\title{A function which makes Sora smooth splines with boot strapped confidence intervals}
\usage{
sora_spline(data, spar = 0.8, year)
}
\arguments{
\item{data}{= a data frame with three columns, in this order, ordinal date, sora per hour and the year}

\item{spar}{= the smoothing parameter, defaults to .8}

\item{year}{= the year the data were collected}
}
\description{
A function which makes Sora smooth splines with boot strapped confidence intervals
}
\examples{
sora_spline()
Needs data with three columns, ordinal date, sora perhour and the year
}
\keyword{rails}

