\name{summary.rel}
\alias{summary.rel}
\title{Class summary structure}
\description{The structure of the summary output for package functions}
\usage{
  \method{summary}{rel}(object, ...)
}
\arguments{
  \item{object}{rel function output}
  \item{...}{Other}
}
\value{The summary output is detailed in the description of the respective functions}