% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fournier.R
\name{theme_fournier}
\alias{theme_fournier}
\title{My custom ggplot2 theme.}
\usage{
theme_fournier()
}
\description{
My custom ggplot2 theme.
}
\examples{
ggplot(data=dat, aes(x=x, y=y))+geom_point()+theme_fournier()
}
\keyword{ggplot2}
