% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/theme_krementz.R
\name{theme_krementz}
\alias{theme_krementz}
\title{A theme that Krementz seems to not hate too much}
\usage{
theme_krementz()
}
\description{
THis function makes ggplots look in a way that my adviser will hopefully not hate
there are no arguments
}
\examples{
ggplot(data=dat, aes(x=x, y=y))+geom_point()+theme_krementz()
}
\keyword{ggplot2}

