% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rallidaerule.R
\name{theme_rallidaerule}
\alias{theme_rallidaerule}
\title{A custom ggplot2 theme to make plots the way I generally like them}
\usage{
theme_rallidaerule()
}
\description{
there are no arguments
}
\examples{
ggplot(data=dat, aes(x=x, y=y))+
     geom_point()+
     theme_rallidaerule()
}
\keyword{ggplot2}
