% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mround.R
\name{mround}
\alias{mround}
\title{Rounds numbers into categories by dec}
\usage{
mround(x, dec)
}
\arguments{
\item{a}{number that you want to round to a dec (decimal)}
}
\description{
Rounds numbers into categories by dec
}
\examples{
mround()
}
\keyword{rounding}

