% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rails_function.R
\name{rail_function}
\alias{rail_function}
\title{A Rails Function}
\usage{
rail_function(love = TRUE)
}
\arguments{
\item{love}{Do you love rails? Defaults to TRUE.}
}
\description{
This function allows you to express your love of rails.
}
\examples{
rails_function()
}
\keyword{rails}

