\name{rel-package}
\alias{rel}
\alias{rel-package}
\docType{package}
\title{Reliability coefficients}
\description{
  \code{rel} provides functions for evaluating agreement between measurements: Bennett et als S (\code{\link{bS}}), Cohen's kappa, weighted kappa, and Conger's kappa (\code{\link{cKap}}), Gwet's AC1 and AC2 (\code{\link{gAC}}), Krippendorff's alpha (\code{\link{krA}}), and Scott's pi and Fleiss' kappa (\code{\link{sPi}}).
}
\keyword{ 
  package
  univar
}