\name{rel-package}
\alias{rel}
\alias{rel-package}
\docType{package}
\title{Reliability coefficients}
\description{
  \code{rel} provides functions for evaluating agreement between measurements: Bennett et als S (\code{\link{bags}}), Cohen's kappa, weighted kappa, and Conger's kappa (\code{\link{ckap}}), Gwet's AC1 and AC2 (\code{\link{gac}}), intraclass correlation coefficients (\code{\link{icc}}), Krippendorff's alpha (\code{\link{kra}}), Scott's pi and Fleiss' kappa (\code{\link{spi}}), and the standard error of measurement (\code{\link{sem}}).
}
\keyword{ 
  package
  univar
}